add changes typed the project > ts
