+++ src/app/components/header/header.component.html
@@ -6,7 +6,8 @@
     <nav class="header-nav">
       <a class="header-nav-item" href="#" routerLink="/">{{header}}</a>
       <a class="header-nav-item" href="#" routerLink="user">{{userButton}}</a>
-      <a class="header-nav-item" href="#">{{aboutCompany}}</a>
+      <a class="header-nav-item" href="#" routerLink="todos">{{todosButton}}</a>
+      <a class="header-nav-item" *ngIf="isShowCatalog" href="#">{{aboutCompany}}</a>
       <a class="header-nav-item" *ngIf="isShowCatalog" href="#">{{header3}}</a>
     </nav>
     <button (click)="isShowCatalog=!isShowCatalog">скрыть/паказать каталог</button>
