+++ src/app/todos-list/todos-list.component.scss
@@ -14,3 +14,9 @@
   padding: 16px;
   box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.5);
 }
+.todo-lists {
+  text-align: center;
+  font-size: 24px;
+  font-weight: bold;
+  color: #333;
+}
