+++ src/app/todos-list/todos-list.component.ts
@@ -2,13 +2,8 @@ import { Component, inject } from '@angular/core';
 import { NgFor, NgIf } from "@angular/common";
 import { TodoCardComponent } from "./todos-card/todos-card.component";
 import { TodosApiService } from '../todos-api.service';
+import { Todo } from "../models/todo.model";
 
-export interface Todo{
-  userId:number;
-  id:number;
-  title:string;
-  completed:boolean;
-}
 
 @Component({
   selector: 'app-todos-list',
@@ -21,16 +16,14 @@ export interface Todo{
 export class TodosListComponent {
   readonly todosApiService = inject(TodosApiService);
   todos: Todo[] = [];
-
   constructor() {
     this.todosApiService.getTodos().subscribe(
-      (response: any) => {
+      (response: Todo[]) => {
         this.todos = response;
-      }
-    );
+      });
   }
 
-  deleteTodo(id: any) {
+  deleteTodo(id: number) {
     this.todos = this.todos.filter(todo => todo.id !== id);
   }
 }
