+++ src/app/user-list/user-card/user-card.component.html
@@ -2,4 +2,4 @@
 <p>{{ user.email }}</p>
 <p>{{ user.website }}</p>
 <p>{{ user.company.name }}</p>
-<button (click)="onDeleteUser(user.id)" class="delete-button">удалить user</button>
+<button (click)="onDeleteUser()" class="delete-button">Удалить пользователя</button>
