+++ src/app/user-list/user-list.component.ts
@@ -4,30 +4,8 @@ import { HeaderComponent } from '../components/header/header.component';
 import { HttpClient } from "@angular/common/http";
 import { UsersApiService } from '../user-api.service';
 import { UserCardComponent } from "./user-card/user-card.component";
+import { User } from "../models/user.model";
 
-export interface User {
-  id: number;
-  name: string;
-  username: string;
-  email: string;
-  address: {
-    street: string;
-    suite: string;
-    city: string;
-    zipcode: string;
-    geo: {
-      lat: string;
-      lng: string;
-    };
-  };
-  phone: string;
-  website: string;
-  company: {
-    name: string;
-    catchPhrase: string;
-    bs: string;
-  };
-}
 
 @Component({
   selector: 'app-user-list',
@@ -42,7 +20,7 @@ export class UserListComponent {
   users: User[] = [];
   constructor(){
     this.usersApiService.getUsers()
-      .subscribe((response: any) => {
+      .subscribe((response: User[]) => {
         this.users = response;
       });
   }
@@ -51,3 +29,5 @@ export class UserListComponent {
     this.users = this.users.filter(user => user.id !== id);
   }
 }
+export { User };
+
